Linted project

+++ App.js
@@ -1,10 +1,15 @@
 import React from 'react';
-import { StyleSheet, Text, View, AsyncStorage } from 'react-native';
+import {
+  Text,
+  View,
+  AsyncStorage,
+  Alert,
+} from 'react-native';
 import styles from './styles/App-styles';
 import Main from './components/Main';
 
 export default class App extends React.Component {
-  constructor(){
+  constructor() {
     super()
     this.state = {
       items: [],
@@ -19,40 +24,50 @@ export default class App extends React.Component {
           AsyncStorage.setItem('items', JSON.stringify([]))
           return
         }
-        return parsedItems
+        return parsedItems // eslint-disable-line
       })
-      .then((parsedItems) => {this.setState({ items: parsedItems })})
+      .then((parsedItems) => { this.setState({ items: parsedItems }) })
       .catch((err) => { throw new Error(err) })
   }
 
-  addNewItem(newItem){
-   this.setState({ items: [
-     ...this.state.items,
-     newItem
-   ] });
-   AsyncStorage.setItem('items', JSON.stringify([
-     ...this.state.items,
-     newItem
-   ]))
- }
+  addNewItem = (newItem) => {
+    this.setState({ items: [
+      ...this.state.items,
+      newItem,
+    ] });
+    AsyncStorage.setItem('items', JSON.stringify([
+      ...this.state.items,
+      newItem,
+    ]))
+  }
 
- deleteAllItems(){
-   this.setState({ items: [] });
- }
+  deleteAllItems = () => {
+    this.setState({ items: [] });
+  }
 
- deleteItem = (id) => {
-   let newArr = this.state.items.filter((item) => {
-       return item.id !== id
-     })
-     AsyncStorage.setItem('items', JSON.stringify(
-       newArr
-     ))
-     .then(() => {this.setState({ items: newArr })})
+  deleteItem = (id) => {
+    Alert.alert(
+     'Warning',
+     'You are about to delete this item! This cannot be undone!',
+      [
+        {
+          text: 'OK',
+          onPress: () => {
+            const newArr = this.state.items.filter((item) => {
+              return item.id !== id
+            })
+            AsyncStorage.setItem('items', JSON.stringify(
+               newArr,
+             ))
+             .then(() => { this.setState({ items: newArr }) })
+          },
+        },
+      ],
+   )
+  }
 
- }
-
- sortAlpha = () => {
-    let newArr = this.state.items.sort((a, b) => {
+  sortAlpha = () => {
+    const newArr = this.state.items.sort((a, b) => {
       const first = a.name.toLowerCase()
       const second = b.name.toLowerCase()
       if (first < second) {
@@ -62,24 +77,29 @@ export default class App extends React.Component {
         return 1
       }
       return 0
-      }
+    },
     );
     this.setState({ items: newArr });
   }
 
-  sortByAisle(){
-    let newArr = this.state.items.sort((a, b) => {return a.aisle - b.aisle });
+  sortByAisle = () => {
+    const newArr = this.state.items.sort((a, b) => { return a.aisle - b.aisle });
     this.setState({ items: newArr });
   }
 
   render() {
-     const { items } = this.state
+    const { items } = this.state
     return (
       <View style={styles.container}>
         <Text style={styles.text}>
           Flash Shopper
         </Text>
-        <Text>{items.length ? <Text>You have {items.length} items on your list.</Text> : <Text>There are no items on your list!</Text>}</Text>
+        <Text>
+          {items.length ?
+            <Text>You have {items.length} items on your list.</Text>
+            :
+            <Text>There are no items on your list!</Text>}
+        </Text>
         <Main
           addNewItem={this.addNewItem.bind(this)}
           items={items}
@@ -87,8 +107,7 @@ export default class App extends React.Component {
           sortByAisle={this.sortByAisle.bind(this)}
           sortAlpha={this.sortAlpha.bind(this)}
           deleteAllItems={this.deleteAllItems.bind(this)}
-          >
-        </Main>
+        />
       </View>
     );
   }
